# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API documentation for Printart Accurate",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
...
        "/account-count": {
            "get": {
                "tags": [
                    "GLAccount"
                ],
                "summary": "Count GL Account Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/item-count": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Count Item Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/transaction/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Find Transaction by Invoice Number",
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice Number of transaction that needs to be fetched",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "PURCHASEORDERNO": {
                                            "type": "string"
                                        },
                                        "ARINVOICEID": {
                                            "type": "string"
                                        },
                                        "CUSTOMERID": {
                                            "type": "number"
                                        },
                                        "SALESMANID": {
                                            "type": "number"
                                        },
                                        "INVOICENO": {
                                            "type": "string"
                                        },
                                        "WAREHOUSEID": {
                                            "type": "number"
                                        },
                                        "INVOICEDATE": {
                                            "type": "string"
                                        },
                                        "INVOICEAMOUNT": {
                                            "type": "number"
                                        },
                                        "PAIDAMOUNT": {
                                            "type": "number"
                                        },
                                        "RATE": {
                                            "type": "string"
                                        },
                                        "TERMDISCOUNT": {
                                            "type": "number"
                                        },
                                        "RETURNAMOUNT": {
                                            "type": "number"
                                        },
                                        "OWING": {
                                            "type": "number"
                                        },
                                        "TERMSID": {
                                            "type": "number"
                                        },
                                        "GLPERIOD": {
                                            "type": "string"
                                        },
                                        "GLYEAR": {
                                            "type": "string"
                                        },
                                        "PRINTED": {
                                            "type": "string"
                                        },
                                        "SHIPDATE": {
                                            "type": "string"
                                        },
                                        "TAX1RATE": {
                                            "type": "string"
                                        },
                                        "TAX2RATE": {
                                            "type": "string"
                                        },
                                        "GLHISTID": {
                                            "type": "number"
                                        },
                                        "PAYMENT": {
                                            "type": "string"
                                        },
                                        "CASHDISCOUNT": {
                                            "type": "string"
                                        },
                                        "TEMPLATEID": {
                                            "type": "string"
                                        },
                                        "ARACCOUNT": {
                                            "type": "string"
                                        },
                                        "GETFROMOTHER": {
                                            "type": "string"
                                        },
                                        "DELIVERYORDER": {
                                            "type": "string"
                                        },
                                        "GETFROMSO": {
                                            "type": "string"
                                        },
                                        "TRANSACTIONID": {
                                            "type": "string"
                                        },
                                        "FREIGHT": {
                                            "type": "string"
                                        },
                                        "ISOUTSTANDING": {
                                            "type": "string"
                                        },
                                        "OUTSTANDINGDO": {
                                            "type": "string"
                                        },
                                        "DPAMOUNT": {
                                            "type": "string"
                                        },
                                        "DPTAX": {
                                            "type": "string"
                                        },
                                        "PROJECTAMOUNT": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transaction-count": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Count Transaction Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}